pyproject.toml
[tool.poetry]
name = "course-project-5"
version = "0.1.0"
description = ""
authors = ["ArtemMTim <artemtimofeev@rambler.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pillow = "^11.0.0"
psycopg2-binary = "^2.9.10"
python-dotenv = "^1.0.1"
django = "^5.1.4"
django-filter = "^24.3"
djangorestframework = "^3.15.2"
djangorestframework-simplejwt = "^5.3.1"
requests = "^2.32.3"
coverage = "^7.6.9"


[tool.poetry.group.lint.dependencies]
flake8 = "^7.1.1"
black = "^24.10.0"
isort = "^5.13.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
